Fixes and enhancement to base journal reading
EDCommander on creation gets EDDN ticked and does not get path in EDSM API entry
EDJournalClass renamed, moved to EDScanner.cs to make it clearer its the top level scanner.
EDMonitorWatcher improved so it can read and fire back N files from the end of play.
EDJournalReader can handle continued if in play and it does not have a DB. Bug fixed so the commander ID is set correctly on the play over of continued.

JournalReaderEntry is removed - now use JsonCached in JE to carry JSON.  JsonCached will be unpopulated for entrys from the DB but set up for new entries read during play