Speed up TLU read
Missed file.
taking too long to start eddlite on HD
TLU has a cache which gets autofilled on first read then the cache is used from then on.
Monitorwatcher removed its TLU cache, and when it needs to add readers to the TLU it stores them in a list and adds in one transaction - thus speeding up the initial scan completely