Major fix to journal, TLU, netlog scanners
Journal TLU system was only using the filename, not the full path, as the key. This was crashing the program if you have two logs with the same filename, as I do.  Change TLU and systems to key off full path.
TLU/TLUReader sanitised for variable names.
EDMonitorWatcher uses full file name for the key.
Journal reader throwing on error is a VERY bad idea, as it stops the rest of the file being scanned for valid data.  Now gracefully says Null, using the new JSON system, but means the scanner continues with the rest of the file.  Not continuing meant that later, in normal use, the normal stuff was scanning it and waiting around for non existant files to be produced (ReadAdditionalFiles)
Added journalcargotrasfer MissionID to transfer list
Added more drone types to DroneType
Added Name_Localised to EngineeringMods
Added Name_Localised to tech broker
Netlog scanner was very broken, causing DB issues with reentrancy. Fixed. Recoded to match better the journal scanner.
ShipModule - added ValueStr_Localised, and uses new Json toobject property of being able to ignore type errors and continue processing the object so we can at least get most of the data present in those broken journal entries.